requirements.txt
streamlit
bokeh
pandas
altair
plotly
numpy
matplotlib
sentence_transformers
umap_learn